Use non-compliant URL from 'url' for socks urls
`new URL('socks://host:port')` gives different results in browser and
Node.js because `socks` is not among the "special schemes" specified by
WHATWG and thus the hostname is treated as an opaque value (without
parsing the port). Node.js implementation, however, parses the `port`.

Since `proxy-agent` can be used in Electron, unless `URL` is imported
directly from the Node's `url` module - it is going to use the Browser's
version of it which won't work with socks urls.