Make decoding of base64 data URIs faster
I saw in https://github.com/microsoft/vscode-js-debug/issues/1911 that
base64 decoding of a data URI was taking a bit of time.

This PR feature-detects the presence of a global `Buffer` to use native
decoding when running in Node.js, which is about 25x faster on a 10MB
data URI than the JS implementation in the library.

I have a bit of a hack in order to test both paths when running tests,
happy to change it if desired :)